Macierz generowana przez POST i tlumaczenie pluginu, brakuje tylko w niej uzupełnienia ^Cla z nazwą projektu lub filtra

+++ src/main/java/net/amg/jira/plugins/jrmp/services/ImpactProbability.java
@@ -21,5 +21,5 @@ import com.atlassian.query.Query;
  */
 public interface ImpactProbability {
 
-    double getMaxProbability(Query query);
+    int getMaxProbability(Query query);
 }
